requirements.txt
fire==0.5.0
loguru==0.6.0
pyyaml==6.0
openpyxl==3.1.1